~ refactoring syntax_errors.ts
