+++ server/src/features/checks/checkInfixOperators.ts
@@ -1,6 +1,6 @@
 import Parser = require("web-tree-sitter");
 import { DeclarationType, TamarinSymbolTable, TamarinSymbol} from '../../symbol_table/create_symbol_table';
-import { getName } from '../syntax_errors';
+import { getName } from './utils';
 import { TextDocument } from "vscode-languageserver-textdocument";
 import { Diagnostic , DiagnosticSeverity ,Range} from 'vscode-languageserver';
